move RaftTestM to new file, begin making RaftTestT
